refactor: IO operations and improve error handling
- Implement read_exact for Read trait and move impl blocks to io_impl.rs
- Remove utils.rs and adjust related imports
