+++ crates/ckb-script-ipc-common/src/lib.rs
@@ -10,5 +10,4 @@ pub mod ipc;
 pub mod packet;
 pub mod pipe;
 pub mod spawn;
-pub mod utils;
 pub mod vlq;
